Moving functor filter to experimental
This will prevent bad API haunting PCL in the future
